samples: add pushing to internal reg + docker-image
Previously, we didn't use to push the images built by estalero to a
registry. With the addition of a custom one to hush-house, now we can do
that regardless of whether we use OCI mediatypes or not (giving
flexibility for us to try out OCI image layouts in the future).

For now, standard images are being pushed.

PS.: `docker-image` sample is not ready yet as we need to add back the
ability to include custom gpg repository keys

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>