cmd/frontend: initial support for apt key
Here we start adding support for allowing one to specify additional apt
keys that should be trusted by `apt` when retrieving package indices
from debian repositories.

```
image "img" {
	apt {
		key "blabla" {
			uri = "https://this.com/gpg"
		}
	}
}
```

As to do so, we also need to tell the bill of materials what were those
keys trusted, we include the ability to do so through `estaleiro
apt-key`, which is able to tell details about those public keys added.

---

As the final retrieval relies on `opengpg` being installed, we do so
"off-the-tree", just giving back to the final `fs` tree the
`trusted.gpg` that gets generated.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>