fe/cmd: support for https-based repositories
With the needs of having the retrieval packages from HTTPS-based
repositories, we end up adding `ca-certificates` to the mix of
dependencies for the repositories retrieval step.

That, however, showed that I have messed up with the dependency
calculation that `apt` preforms when I put both package and repository
retrieval in the same vertex - having `ca-certificates` for retrieving
https repositories informations would just make `ca-certificates` be
marked as an already "installed" package later on.

To solve that, I end up splitting that command in two (apt_repositories
and apt_packages), mounting the necessary things to the second (so that
we can still leverage ca-certificates), while still not polluting its
dependency tree.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>